Resolves: #10094

*(short description of the changes and the motivation to make the
changes)*

<!-- Use "x" to fill the checkboxes below like [x] -->

- [ ] I signed [CLA](https://www.audacityteam.org/cla/)
- [ ] The title of the pull request describes an issue it addresses
- [ ] If changes are extensive, then there is a sequence of easily
reviewable commits
- [ ] Each commit's message describes its purpose and effects
- [ ] There are no behavior changes unnecessary for the stated purpose
of the PR

Recommended:
- [ ] Each commit compiles and runs on my machine without known
undesirable changes of behavior

QA:

- [ ] Autobot test cases have been run
- [x] Clip sample editing still works for both mono and stereo clips

+++ src/spectrogram/internal/au3/SpectrumCache.h
@@ -62,8 +62,8 @@ public:
     double rightTrim{ .0 };
     double start;      // relative to clip start
     std::optional<SpectrogramWindowType> windowType;
-    size_t windowSize { 0 };
-    unsigned zeroPaddingFactor { 0 };
+    int windowSize { 0 };
+    int zeroPaddingFactor { 0 };
     int frequencyGain;
     std::vector<float> freq;
     std::vector<long long> where;
